+++ src/de/map.rs
@@ -2,18 +2,23 @@
 
 use crate::{
     de::escape::EscapedDeserializer,
-    de::{BorrowingReader, Deserializer, INNER_VALUE, UNFLATTEN_PREFIX},
+    de::{BorrowingReader, DeEvent, Deserializer, INNER_VALUE, UNFLATTEN_PREFIX},
     errors::serialize::DeError,
-    events::{BytesStart, Event},
+    events::attributes::Attribute,
+    events::BytesStart,
 };
 use serde::de::{self, DeserializeSeed, IntoDeserializer};
+use std::borrow::Cow;
 
-enum MapValue {
+/// Representing state of the `MapAccess` accessor.
+enum State {
+    /// `next_key_seed` not yet called. This is initial state and state after deserializing
+    /// value (calling `next_value_seed`).
     Empty,
-    /// Value should be deserialized from the attribute value
-    Attribute {
-        value: Vec<u8>,
-    },
+    /// `next_key_seed` checked the attributes list and find it is not exhausted yet.
+    /// Next call to the `next_value_seed` will deserialize type from the attribute value
+    Attribute,
+    /// The same as `InnerValue`
     Nested,
     /// Value should be deserialized from the text content of the XML node:
     ///
@@ -24,7 +29,7 @@ enum MapValue {
 }
 
 /// A deserializer for `Attributes`
-pub(crate) struct MapAccess<'de, 'a, R: BorrowingReader<'de> + 'a> {
+pub(crate) struct MapAccess<'de, 'a, R: BorrowingReader<'de>> {
     /// Tag -- owner of attributes
     start: BytesStart<'de>,
     de: &'a mut Deserializer<'de, R>,
@@ -34,9 +39,9 @@ pub(crate) struct MapAccess<'de, 'a, R: BorrowingReader<'de> + 'a> {
     /// a new object on each advance of `Attributes` iterator, so we need
     /// to restore last position before advance.
     position: usize,
-    value: MapValue,
-    /// number of fields yet to parse
-    size_hint: Option<usize>,
+    /// Current state of the accessor that determines what next call to API
+    /// methods should return.
+    state: State,
     /// list of fields yet to unflatten (defined as starting with $unflatten=)
     unflatten_fields: Vec<&'static [u8]>,
 }
@@ -53,12 +58,7 @@ impl<'de, 'a, R: BorrowingReader<'de>> MapAccess<'de, 'a, R> {
             de,
             start,
             position,
-            value: MapValue::Empty,
-            size_hint: if fields.is_empty() {
-                None
-            } else {
-                Some(fields.len())
-            },
+            state: State::Empty,
             unflatten_fields: fields
                 .iter()
                 .filter(|f| f.starts_with(UNFLATTEN_PREFIX))
@@ -67,21 +67,17 @@ impl<'de, 'a, R: BorrowingReader<'de>> MapAccess<'de, 'a, R> {
         })
     }
 
-    fn next_attr(&mut self) -> Result<Option<(Vec<u8>, Vec<u8>)>, DeError> {
+    fn next_attr(&mut self) -> Result<Option<Attribute>, DeError> {
         let mut attributes = self.start.attributes();
         attributes.position = self.position;
         let next_att = attributes.next().transpose()?;
         self.position = attributes.position;
-        Ok(next_att.map(|a| (a.key.to_owned(), a.value.into_owned())))
+        Ok(next_att)
     }
 }
 
-impl<'de, 'a, R: BorrowingReader<'de> + 'a> de::MapAccess<'de> for MapAccess<'de, 'a, R> {
+impl<'de, 'a, R: BorrowingReader<'de>> de::MapAccess<'de> for MapAccess<'de, 'a, R> {
     type Error = DeError;
-
-    fn size_hint(&self) -> Option<usize> {
-        self.size_hint.clone()
-    }
 
     fn next_key_seed<K: DeserializeSeed<'de>>(
         &mut self,
@@ -89,17 +85,23 @@ impl<'de, 'a, R: BorrowingReader<'de> + 'a> de::MapAccess<'de> for MapAccess<'de
     ) -> Result<Option<K::Value>, Self::Error> {
         let decoder = self.de.reader.decoder();
         let has_value_field = self.de.has_value_field;
-        if let Some((key, value)) = self.next_attr()? {
+
+        let mut attributes = self.start.attributes();
+        attributes.position = self.position;
+        if let Some(a) = attributes.next().transpose()? {
             // try getting map from attributes (key= "value")
-            self.value = MapValue::Attribute { value };
-            self.size_hint.as_mut().map(|l| *l = l.wrapping_sub(1));
-            seed.deserialize(EscapedDeserializer::new(key, decoder, false))
-                .map(Some)
+            self.state = State::Attribute;
+            seed.deserialize(EscapedDeserializer::new(
+                Cow::Borrowed(a.key),
+                decoder,
+                false,
+            ))
+            .map(Some)
         } else {
             // try getting from events (<key>value</key>)
             match self.de.peek()? {
-                Some(Event::Text(_)) => {
-                    self.value = MapValue::InnerValue;
+                DeEvent::Text(_) => {
+                    self.state = State::InnerValue;
                     // Deserialize `key` from special attribute name which means
                     // that value should be taken from the text content of the
                     // XML node
@@ -121,13 +123,11 @@ impl<'de, 'a, R: BorrowingReader<'de> + 'a> de::MapAccess<'de> for MapAccess<'de
                 // }
                 // TODO: This should be handled by #[serde(flatten)]
                 // See https://github.com/serde-rs/serde/issues/1905
-                Some(Event::Start(_)) if has_value_field => {
-                    self.value = MapValue::InnerValue;
-                    self.size_hint.as_mut().map(|l| *l = l.wrapping_sub(1));
+                DeEvent::Start(_) if has_value_field => {
+                    self.state = State::InnerValue;
                     seed.deserialize(INNER_VALUE.into_deserializer()).map(Some)
                 }
-                Some(Event::Start(e)) => {
-                    self.size_hint.as_mut().map(|l| *l = l.wrapping_sub(1));
+                DeEvent::Start(e) => {
                     let key = if let Some(p) = self
                         .unflatten_fields
                         .iter()
@@ -144,11 +144,11 @@ impl<'de, 'a, R: BorrowingReader<'de> + 'a> de::MapAccess<'de> for MapAccess<'de
                         //     #[serde(rename = "$unflatten=xxx")]
                         //     xxx: String,
                         // }
-                        self.value = MapValue::InnerValue;
+                        self.state = State::InnerValue;
                         seed.deserialize(self.unflatten_fields.remove(p).into_deserializer())
                     } else {
-                        let name = e.local_name().to_owned();
-                        self.value = MapValue::Nested;
+                        let name = Cow::Borrowed(e.local_name());
+                        self.state = State::Nested;
                         seed.deserialize(EscapedDeserializer::new(name, decoder, false))
                     };
                     key.map(Some)
@@ -162,14 +162,17 @@ impl<'de, 'a, R: BorrowingReader<'de> + 'a> de::MapAccess<'de> for MapAccess<'de
         &mut self,
         seed: K,
     ) -> Result<K::Value, Self::Error> {
-        match std::mem::replace(&mut self.value, MapValue::Empty) {
-            MapValue::Attribute { value } => seed.deserialize(EscapedDeserializer::new(
-                value,
-                self.de.reader.decoder(),
-                true,
-            )),
-            MapValue::Nested | MapValue::InnerValue => seed.deserialize(&mut *self.de),
-            MapValue::Empty => Err(DeError::EndOfAttributes),
+        match std::mem::replace(&mut self.state, State::Empty) {
+            State::Attribute => {
+                let decoder = self.de.reader.decoder();
+                match self.next_attr()? {
+                    Some(a) => seed.deserialize(EscapedDeserializer::new(a.value, decoder, true)),
+                    // We set `Attribute` state only when we are sure that `next_attr()` returns a value
+                    None => unreachable!(),
+                }
+            }
+            State::Nested | State::InnerValue => seed.deserialize(&mut *self.de),
+            State::Empty => Err(DeError::EndOfAttributes),
         }
     }
 }