+++ src/de/seq.rs
@@ -1,9 +1,6 @@
-use crate::de::{BorrowingReader, DeError, Deserializer};
-use crate::{
-    events::{BytesStart, Event},
-    reader::Decoder,
-};
-use serde::de;
+use crate::de::{BorrowingReader, DeError, DeEvent, Deserializer};
+use crate::{events::BytesStart, reader::Decoder};
+use serde::de::{self, DeserializeSeed};
 
 #[derive(Debug)]
 enum Names {
@@ -26,20 +23,25 @@ impl Names {
 }
 
 /// A SeqAccess
-pub struct SeqAccess<'de, 'a, R: BorrowingReader<'de>> {
+pub struct SeqAccess<'de, 'a, R>
+where
+    R: BorrowingReader<'de>,
+{
     de: &'a mut Deserializer<'de, R>,
-    max_size: Option<usize>,
     names: Names,
 }
 
-impl<'a, 'de, R: BorrowingReader<'de>> SeqAccess<'de, 'a, R> {
+impl<'a, 'de, R> SeqAccess<'de, 'a, R>
+where
+    R: BorrowingReader<'de>,
+{
     /// Get a new SeqAccess
-    pub fn new(de: &'a mut Deserializer<'de, R>, max_size: Option<usize>) -> Result<Self, DeError> {
+    pub fn new(de: &'a mut Deserializer<'de, R>) -> Result<Self, DeError> {
         let decoder = de.reader.decoder();
         let names = if de.has_value_field {
             Names::Unknown
         } else {
-            if let Some(Event::Start(e)) = de.peek()? {
+            if let DeEvent::Start(e) = de.peek()? {
                 #[cfg(not(feature = "encoding"))]
                 let name = decoder.decode(e.name())?.to_owned();
                 #[cfg(feature = "encoding")]
@@ -49,35 +51,24 @@ impl<'a, 'de, R: BorrowingReader<'de>> SeqAccess<'de, 'a, R> {
                 Names::Unknown
             }
         };
-        Ok(SeqAccess {
-            de,
-            max_size,
-            names,
-        })
+        Ok(SeqAccess { de, names })
     }
 }
 
-impl<'de, 'a, R: BorrowingReader<'de>> de::SeqAccess<'de> for SeqAccess<'de, 'a, R> {
+impl<'de, 'a, R> de::SeqAccess<'de> for SeqAccess<'de, 'a, R>
+where
+    R: BorrowingReader<'de>,
+{
     type Error = DeError;
 
-    fn size_hint(&self) -> Option<usize> {
-        self.max_size
-    }
-
-    fn next_element_seed<T: de::DeserializeSeed<'de>>(
-        &mut self,
-        seed: T,
-    ) -> Result<Option<T::Value>, DeError> {
-        if let Some(s) = self.max_size.as_mut() {
-            if *s == 0 {
-                return Ok(None);
-            }
-            *s -= 1;
-        }
+    fn next_element_seed<T>(&mut self, seed: T) -> Result<Option<T::Value>, DeError>
+    where
+        T: DeserializeSeed<'de>,
+    {
         let decoder = self.de.reader.decoder();
         match self.de.peek()? {
-            None | Some(Event::Eof) | Some(Event::End(_)) => Ok(None),
-            Some(Event::Start(e)) if !self.names.is_valid(decoder, e)? => Ok(None),
+            DeEvent::Eof | DeEvent::End(_) => Ok(None),
+            DeEvent::Start(e) if !self.names.is_valid(decoder, e)? => Ok(None),
             _ => seed.deserialize(&mut *self.de).map(Some),
         }
     }