+++ src/de/var.rs
@@ -1,76 +1,96 @@
 use crate::{
-    de::{escape::EscapedDeserializer, BorrowingReader, Deserializer},
+    de::{escape::EscapedDeserializer, BorrowingReader, DeEvent, Deserializer},
     errors::serialize::DeError,
-    events::Event,
 };
-use serde::de::{self, Deserializer as SerdeDeserializer};
+use serde::de::{self, DeserializeSeed, Deserializer as SerdeDeserializer, Visitor};
+use std::borrow::Cow;
 
 /// An enum access
-pub struct EnumAccess<'de, 'a, R: BorrowingReader<'de>> {
+pub struct EnumAccess<'de, 'a, R>
+where
+    R: BorrowingReader<'de>,
+{
     de: &'a mut Deserializer<'de, R>,
 }
 
-impl<'de, 'a, R: BorrowingReader<'de>> EnumAccess<'de, 'a, R> {
+impl<'de, 'a, R> EnumAccess<'de, 'a, R>
+where
+    R: BorrowingReader<'de>,
+{
     pub fn new(de: &'a mut Deserializer<'de, R>) -> Self {
         EnumAccess { de }
     }
 }
 
-impl<'de, 'a, R: BorrowingReader<'de>> de::EnumAccess<'de> for EnumAccess<'de, 'a, R> {
+impl<'de, 'a, R> de::EnumAccess<'de> for EnumAccess<'de, 'a, R>
+where
+    R: BorrowingReader<'de>,
+{
     type Error = DeError;
     type Variant = VariantAccess<'de, 'a, R>;
 
-    fn variant_seed<V: de::DeserializeSeed<'de>>(
-        self,
-        seed: V,
-    ) -> Result<(V::Value, VariantAccess<'de, 'a, R>), DeError> {
+    fn variant_seed<V>(self, seed: V) -> Result<(V::Value, VariantAccess<'de, 'a, R>), DeError>
+    where
+        V: DeserializeSeed<'de>,
+    {
         let decoder = self.de.reader.decoder();
         let de = match self.de.peek()? {
-            Some(Event::Text(t)) => EscapedDeserializer::new(t.to_vec(), decoder, true),
-            Some(Event::Start(e)) => EscapedDeserializer::new(e.name().to_vec(), decoder, false),
-            Some(e) => return Err(DeError::InvalidEnum(e.clone().into_owned())),
-            None => return Err(DeError::Eof),
+            DeEvent::Text(t) => EscapedDeserializer::new(Cow::Borrowed(t), decoder, true),
+            DeEvent::Start(e) => EscapedDeserializer::new(Cow::Borrowed(e.name()), decoder, false),
+            _ => {
+                return Err(DeError::Unsupported(
+                    "Invalid event for Enum, expecting `Text` or `Start`",
+                ))
+            }
         };
         let name = seed.deserialize(de)?;
         Ok((name, VariantAccess { de: self.de }))
     }
 }
 
-pub struct VariantAccess<'de, 'a, R: BorrowingReader<'de>> {
+pub struct VariantAccess<'de, 'a, R>
+where
+    R: BorrowingReader<'de>,
+{
     de: &'a mut Deserializer<'de, R>,
 }
 
-impl<'de, 'a, R: BorrowingReader<'de>> de::VariantAccess<'de> for VariantAccess<'de, 'a, R> {
+impl<'de, 'a, R> de::VariantAccess<'de> for VariantAccess<'de, 'a, R>
+where
+    R: BorrowingReader<'de>,
+{
     type Error = DeError;
 
     fn unit_variant(self) -> Result<(), DeError> {
         match self.de.next()? {
-            Event::Start(e) => self.de.read_to_end(e.name()),
-            Event::Text(_) => Ok(()),
+            DeEvent::Start(e) => self.de.read_to_end(e.name()),
+            DeEvent::Text(_) => Ok(()),
             _ => unreachable!(),
         }
     }
 
-    fn newtype_variant_seed<T: de::DeserializeSeed<'de>>(
-        self,
-        seed: T,
-    ) -> Result<T::Value, DeError> {
+    fn newtype_variant_seed<T>(self, seed: T) -> Result<T::Value, DeError>
+    where
+        T: DeserializeSeed<'de>,
+    {
         seed.deserialize(&mut *self.de)
     }
 
-    fn tuple_variant<V: de::Visitor<'de>>(
-        self,
-        len: usize,
-        visitor: V,
-    ) -> Result<V::Value, DeError> {
+    fn tuple_variant<V>(self, len: usize, visitor: V) -> Result<V::Value, DeError>
+    where
+        V: Visitor<'de>,
+    {
         self.de.deserialize_tuple(len, visitor)
     }
 
-    fn struct_variant<V: de::Visitor<'de>>(
+    fn struct_variant<V>(
         self,
         fields: &'static [&'static str],
         visitor: V,
-    ) -> Result<V::Value, DeError> {
+    ) -> Result<V::Value, DeError>
+    where
+        V: Visitor<'de>,
+    {
         self.de.deserialize_struct("", fields, visitor)
     }
 }