+++ Changelog.md
@@ -15,6 +15,10 @@
 - refactor: unify errors when EOF encountered during serde deserialization
 - test: ensure that after deserializing all XML was consumed
 - feat: add `Deserializer::from_str` and `Deserializer::from_bytes`
+- refactor: reduce number of unnecessary copies when deserialize numbers/booleans/identifiers
+  from the attribute and element names and attribute values
+- fix: allow to deserialize `unit`s from text and CDATA content.
+  `DeError::InvalidUnit` variant is removed, because after fix it is no longer used
 
 ## 0.23.0-alpha3